pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wikicoder"
version = "0.1.0"
description = "Wiki-first coding CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "prompt-toolkit>=3.0.43",
  "PyYAML>=6.0.1",
  "openai>=1.0.0",
  "httpx>=0.27.0",
]

[project.scripts]
wikicoder = "src.main:run_cli"
wikicoderctl = "src.main:app"

[tool.setuptools]
packages = ["src", "src.core", "src.skills", "src.utils"]
